+++ NEWS.md
@@ -1,3 +1,7 @@
 # reumetsat 0.1.0
 
-* Initial CRAN submission.
+* Read HDF5 data files from AC SAFT FMI server into data frames
+in function `read_AC_SAFT_hdf5()`. These files contain data on a geographic grid.
+* Read text data files from AC SAFT FMI server into data frames
+in function `read_AC_SAFT_txt()`. These files contain a time series of data for
+a single geographic location.